pyproject.toml
[project]
name = "task-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "bcrypt<4.1.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "click==8.3.1",
    "colorama==0.4.6",
    "dnspython==2.8.0",
    "dotenv==0.9.9",
    "fastapi==0.135.1",
    "gevent==25.9.1",
    "greenlet==3.3.2",
    "h11==0.16.0",
    "idna==3.11",
    "motor==3.7.1",
    "passlib==1.7.4",
    "pycparser==3.0",
    "pydantic[email]==2.12.5",
    "pydantic-core==2.41.5",
    "pymongo==4.16.0",
    "python-dotenv==1.2.2",
    "python-multipart==0.0.22",
    "starlette==0.52.1",
    "typing==3.7.4.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "uvicorn[standard]==0.41.0",
    "websocket==0.2.1",
    "zope-event==6.1",
    "zope-interface==8.2",
    "python-jose[cryptography]>=3.5.0",
    "timedelta>=2020.12.3",
    "websockets>=16.0",
    "aiofiles>=25.1.0",
    "redis>=7.4.0",
]
